refactor(gemma4): extract MoE into gemma4_moe.rs; revert QLinear expert matmul
Move Gemma4MoeRouter, Gemma4MoeExperts, Gemma4MoeBlock, MoeExpertWeights,
split_expert_qtensor, and rms_norm_no_scale into a new gemma4_moe.rs module
(~400 lines). gemma4.rs imports the single type it needs via
`use super::gemma4_moe::Gemma4MoeBlock`.

Revert the QLinear expert matmul optimisation: QLinear::from_qtensor with a
CPU-backed QTensor hangs when activations are on Metal because the quantized
GEMV kernel requires both operands on the same device. Restored the working
expert_weight() path (explicit dequantize(device) + matmul). The QLinear
optimisation can be revisited once the QTensors are created on the target
device in split_expert_qtensor.

No logic changes to the working MoE implementation.
